requirements.txt
binance-connector
pandas
certifi
click
binance-futures-connector
diskcache